updated text changes
